Exercise 1.5 complete + bugfix
Because we were using string comparision that would mean that:
9 > 8       is true
9 > 80      is false
"9" > "80"  is true

We need to make sure we're not using strings when doing numeric
comparisons.

Also need to make sure we strip off the newline character when doing a
string comparison. Then comparing the input string with "exit" becomes
easy!