Fire a lambda when a target is reached
This feels like Javascript, even the google result that showed me how to convert
a method to a lambda said I should be using Javascript!

https://www.viget.com/articles/convert-ruby-method-to-lambda/

... I guess I've been doing too much Javascript lately :P

This could lend itself to having the Game concept of collecting multiple goals
before passing the level, etc. while the implementation of World doesn't need to
be concerned by it, it just calls the lambda of the target that was reached.