Prevent unwanted mutation of Robot and Target
When using the OpenStruct version, it is possible to change the properties with
the assignment operator, as demonstrated in the test.

This change replaces the OpenStructs with simple classes that have attr_readers
but no attr_writers. These classes are also frozen for good measure (although
probably not actually required).