Start putting some directory structure in
Each step of the curriculum should be addressed by one (or part of an) exercise.

Ideally, each exercise will have:
- Material for a facilitator to use to explain a topic
- Exercises for the students, with inline comments for instructions
- Solutions for the students to use to catchup if they are pairing, etc.

I'm thinking the class structure should work like this:
- Explain the topic for the day
- Demonstrate on screen
- Students pair up and work on exercises
- Go through solutions at the end of the class

This means that the pair can swap computers and still have the solutions
available to them and it also acts as part of a catchup mechanism, this would
also need some documentation, recorded session or external references to learn
that particular topic.